[JimmyGod] - Added App class - Added static method to access main app instance - Added HelloWindow project to display a window

+++ Framework/Math/Inc/EngineMath.h
@@ -118,7 +118,7 @@ namespace JimmyGod::Math
 
 	inline float Magnitude(const Vector3& v)
 	{
-		return sqrt(MagnitudeSqr(v));
+		return sqrtf(MagnitudeSqr(v));
 	}
 
 	inline Vector3 Normalize(const Vector3& v)
